feat: Extract releases upload-sourcemaps to a separate sourcemaps command
