+++ tests/integration/_cases/help/help.trycmd
@@ -38,6 +38,7 @@ SUBCOMMANDS:
     releases           Manage releases on Sentry.
     repos              Manage repositories on Sentry.
     send-event         Send a manual event to Sentry.
+    sourcemaps         Manage sourcemaps for Sentry releases.
     uninstall          Uninstall the sentry-cli executable.
     upload-proguard    Upload ProGuard mapping files to a project.
